[PDI-7203] Clean output path no longer reports a failure if the output path does not exist.

+++ src/org/pentaho/di/job/entries/hadooptransjobexecutor/JobEntryHadoopTransJobExecutor.java
@@ -903,6 +903,10 @@ public class JobEntryHadoopTransJobExecutor extends JobEntryBase implements Clon
 
   private boolean cleanOutputPath(Configuration conf, Path path) throws IOException {
     FileSystem fs = FileSystem.get(conf);
+    if (!fs.exists(path)) {
+      // If the path does not exist one could think of it as "already cleaned"
+      return true;
+    }
     return fs.delete(path, true);
   }
